+++ packages/ui-components/src/components/Axis/Axis.style.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import { AxisBottom as VxAxisBottom, AxisLeft as VxAxisLeft } from "@visx/axis";
 
 import { styled, theme } from "../../styles";
+import palette from "../../styles/theme/palette";
 import typography from "../../styles/theme/typography";
 
 export type AxisLeftProps = React.ComponentProps<typeof VxAxisLeft> & {
@@ -24,7 +25,7 @@ export type AxisBottomProps = React.ComponentProps<typeof VxAxisBottom> & {
 };
 
 const baseTickLabelProps = {
-  fill: typography.paragraphSmall.color,
+  fill: palette.chart.axisLabel,
   fontFamily: typography.paragraphSmall.fontFamily,
   fontSize: typography.paragraphSmall.fontSize,
   fontWeight: typography.paragraphSmall.fontWeight,
@@ -35,7 +36,7 @@ export const AxisLeft = styled((props: AxisLeftProps) => (
     axisClassName={props.className ?? ""}
     numTicks={props.numTicks ?? 5}
     hideTicks={props.hideTicks}
-    tickStroke={theme.palette.border.default}
+    tickStroke={theme.palette.chart.axis}
     tickLabelProps={() => ({
       textAnchor: "end", // Horizontal anchor
       verticalAnchor: "middle",
@@ -51,7 +52,7 @@ export const AxisBottom = styled((props: AxisBottomProps) => (
   <VxAxisBottom
     axisClassName={props.className ?? ""}
     tickLength={4}
-    tickStroke={theme.palette.border.default}
+    tickStroke={theme.palette.chart.axis}
     tickLabelProps={() => ({
       textAnchor: "middle", // Horizontal anchor
       verticalAnchor: "start",