+++ packages/ui-components/src/components/MetricLegendThreshold/utils.ts
@@ -1,5 +1,5 @@
 import { assert } from "@actnowcoalition/assert";
-import { Metric } from "@actnowcoalition/metrics";
+import { Category, Metric } from "@actnowcoalition/metrics";
 
 export interface CategoryItem {
   /**
@@ -18,15 +18,20 @@ export interface CategoryItem {
    * Formatted value of the current category's upper bound.
    */
   endThreshold?: string;
+  /** Show an arrow indicating the current category */
+  showIndicator: boolean;
 }
 
 /**
  * `getMetricCategoryItems` formats the properties of a given metric's
- * categories into an array of category items (`CategoryItem`) to be
+ * categories into an array of category items (CategoryItem[]) to be
  * passed to MetricLegendThreshold.
  */
 
-export function getMetricCategoryItems(metric: Metric): CategoryItem[] {
+export function getMetricCategoryItems(
+  metric: Metric,
+  currentCategory?: Category
+): CategoryItem[] {
   const metricCategories = metric.categorySet?.categories;
   const metricThresholds = metric.categoryThresholds;
 
@@ -38,5 +43,6 @@ export function getMetricCategoryItems(metric: Metric): CategoryItem[] {
     description: category.description,
     endThreshold:
       metricThresholds && metric.formatValue(metricThresholds[index]),
+    showIndicator: category === currentCategory,
   }));
 }