+++ packages/ui-components/src/components/MetricSeriesChart/MetricSeriesChart.stories.tsx
@@ -6,7 +6,7 @@ import { schemeCategory10 } from "d3-scale-chromatic";
 import { states } from "@actnowcoalition/regions";
 
 import { MetricSeriesChart } from ".";
-import { MetricId, metricCatalog } from "../../stories/mockMetricCatalog";
+import { MetricId } from "../../stories/mockMetricCatalog";
 import { theme } from "../../styles";
 import { SeriesType } from "../SeriesChart";
 
@@ -18,6 +18,11 @@ export default {
 const width = 800;
 const height = 600;
 
+const WY = states.findByRegionIdStrict("56");
+const WA = states.findByRegionIdStrict("56");
+const NY = states.findByRegionIdStrict("36");
+const IN = states.findByRegionIdStrict("18");
+
 const Template: ComponentStory<typeof MetricSeriesChart> = (args) => (
   <MetricSeriesChart {...args} />
 );
@@ -28,20 +33,16 @@ Vaccination.args = {
   height,
   series: [
     {
-      region: states.findByRegionIdStrict("56"),
-      metric: metricCatalog.getMetric(MetricId.MOCK_CASES),
+      region: WY,
+      metric: MetricId.MOCK_CASES,
       type: SeriesType.LINE,
-      lineProps: {
-        stroke: theme.palette.gradient[100],
-      },
+      lineProps: { stroke: theme.palette.gradient[100] },
     },
     {
-      region: states.findByRegionIdStrict("53"),
-      metric: metricCatalog.getMetric(MetricId.MOCK_CASES),
+      region: WA,
+      metric: MetricId.MOCK_CASES,
       type: SeriesType.LINE,
-      lineProps: {
-        stroke: theme.palette.gradient[300],
-      },
+      lineProps: { stroke: theme.palette.gradient[300] },
     },
   ],
 };
@@ -53,8 +54,8 @@ WithNegativeYValues.args = {
   minValue: -10,
   series: [
     {
-      region: states.findByRegionIdStrict("56"),
-      metric: metricCatalog.getMetric(MetricId.MOCK_CASES),
+      region: WY,
+      metric: MetricId.MOCK_CASES,
       type: SeriesType.LINE,
     },
   ],
@@ -66,17 +67,35 @@ TrendsSingleLocation.args = {
   height,
   series: [
     {
-      region: states.findByRegionIdStrict("53"),
-      metric: metricCatalog.getMetric(MetricId.MOCK_CASES),
+      region: WA,
+      metric: MetricId.MOCK_CASES,
       type: SeriesType.BAR,
     },
     {
-      region: states.findByRegionIdStrict("53"),
-      metric: metricCatalog.getMetric(MetricId.MOCK_CASES),
+      region: WA,
+      metric: MetricId.MOCK_CASES,
       type: SeriesType.LINE,
-      lineProps: {
-        stroke: "black",
-      },
+      lineProps: { stroke: "black" },
+    },
+  ],
+};
+
+export const FilteredByDate = Template.bind({});
+FilteredByDate.args = {
+  width,
+  height,
+  dateRange: { startAt: new Date("2022-10-01"), endAt: new Date("2022-11-30") },
+  series: [
+    {
+      region: WA,
+      metric: MetricId.MOCK_CASES,
+      type: SeriesType.BAR,
+    },
+    {
+      region: WA,
+      metric: MetricId.MOCK_CASES,
+      type: SeriesType.LINE,
+      lineProps: { stroke: "black" },
     },
   ],
 };
@@ -87,25 +106,52 @@ TrendsMultipleLocations.args = {
   height,
   series: [
     {
-      region: states.findByRegionIdStrict("36"),
-      metric: metricCatalog.getMetric(MetricId.MOCK_CASES),
+      region: NY,
+      metric: MetricId.MOCK_CASES,
       type: SeriesType.LINE,
       lineProps: { stroke: schemeCategory10[0] },
     },
     {
-      region: states.findByRegionIdStrict("56"),
-      metric: metricCatalog.getMetric(MetricId.MOCK_CASES),
+      region: WY,
+      metric: MetricId.MOCK_CASES,
       type: SeriesType.LINE,
       lineProps: { stroke: schemeCategory10[1] },
     },
     {
-      region: states.findByRegionIdStrict("18"),
-      metric: metricCatalog.getMetric(MetricId.MOCK_CASES),
+      region: IN,
+      metric: MetricId.MOCK_CASES,
       type: SeriesType.LINE,
-      lineProps: {
-        stroke: schemeCategory10[2],
-        strokeDasharray: "4 4",
-      },
+      lineProps: { stroke: schemeCategory10[2], strokeDasharray: "4 4" },
+    },
+  ],
+};
+
+export const MultipleLocationsWithLabels = Template.bind({});
+MultipleLocationsWithLabels.args = {
+  width,
+  height,
+  marginRight: 100,
+  series: [
+    {
+      region: NY,
+      label: NY.shortName,
+      metric: MetricId.MOCK_CASES,
+      type: SeriesType.LINE,
+      lineProps: { stroke: schemeCategory10[0] },
+    },
+    {
+      region: WY,
+      label: WY.shortName,
+      metric: MetricId.MOCK_CASES,
+      type: SeriesType.LINE,
+      lineProps: { stroke: schemeCategory10[1] },
+    },
+    {
+      region: IN,
+      label: IN.shortName,
+      metric: MetricId.MOCK_CASES,
+      type: SeriesType.LINE,
+      lineProps: { stroke: schemeCategory10[2], strokeDasharray: "4 4" },
     },
   ],
 };
@@ -116,14 +162,14 @@ LoadingDelay.args = {
   height,
   series: [
     {
-      region: states.findByRegionIdStrict("36"),
+      region: NY,
       metric: MetricId.MOCK_CASES_DELAY_1S,
       type: SeriesType.LINE,
       lineProps: { stroke: schemeCategory10[0] },
     },
     {
-      region: states.findByRegionIdStrict("56"),
-      metric: metricCatalog.getMetric(MetricId.MOCK_CASES),
+      region: WY,
+      metric: MetricId.MOCK_CASES,
       type: SeriesType.LINE,
       lineProps: { stroke: schemeCategory10[1] },
     },
@@ -136,14 +182,14 @@ LoadingError.args = {
   height,
   series: [
     {
-      region: states.findByRegionIdStrict("36"),
+      region: NY,
       metric: MetricId.MOCK_CASES_ERROR,
       type: SeriesType.LINE,
       lineProps: { stroke: schemeCategory10[0] },
     },
     {
-      region: states.findByRegionIdStrict("56"),
-      metric: metricCatalog.getMetric(MetricId.MOCK_CASES),
+      region: WY,
+      metric: MetricId.MOCK_CASES,
       type: SeriesType.LINE,
       lineProps: { stroke: schemeCategory10[1] },
     },