+++ packages/ui-components/src/components/Select/Select.tsx
@@ -1,0 +1,52 @@
+import React from "react";
+
+import { MenuItem, TextField, Typography } from "@mui/material";
+
+export interface SelectProps<T> {
+  /** List of dropdown options */
+  options: T[];
+  /** Selected option */
+  selectedOption: T;
+  /** Label for the selection menu */
+  label: string;
+  /** Function to access the option value */
+  getValue: (item: T) => string;
+  /** Function to access the option label */
+  getLabel: (item: T) => string;
+  /** Handler to call when the user selects an option */
+  onSelectOption: (value: string) => void;
+}
+
+/**
+ * Simple dropdown menu that allows to select a single option. The component
+ * is uncontrolled, so the parent needs to manage the state.
+ */
+export const Select = <T,>({
+  label,
+  options,
+  selectedOption,
+  onSelectOption,
+  getValue,
+  getLabel,
+}: SelectProps<T>) => (
+  <TextField
+    select
+    variant="filled"
+    fullWidth
+    label={<Typography variant="paragraphSmall">{label}</Typography>}
+    value={getValue(selectedOption)}
+    onChange={(event: React.ChangeEvent<HTMLInputElement>) =>
+      onSelectOption(event.target.value)
+    }
+    SelectProps={{
+      disableUnderline: true,
+      MenuProps: { style: { maxHeight: 300 } },
+    }}
+  >
+    {options.map((option) => (
+      <MenuItem key={getValue(option)} value={getValue(option)}>
+        <Typography noWrap>{getLabel(option)}</Typography>
+      </MenuItem>
+    ))}
+  </TextField>
+);