+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @actnowcoalition/ui-components
 
+## 0.9.0
+
+### Minor Changes
+
+- ecbedde: simplify various prop interfaces
+- 22cc5b8: Removing sortColumnId and sortDirection from CompareTable
+- a6fb51d: rename renderTooltip prop to getTooltip
+
+### Patch Changes
+
+- b26d872: Reorganize storybook
+- 043296b: Fix #421: Render placeholder text in MetricOverview while waiting for data to load to avoid layout shift.
+- 9a483b3: Fix #422: Make MetricMultiProgressBar render Skeleton while data is loading to avoid layout shift.
+- e2d92b8: Add default metric format options and decimal points mock data
+- fbc8a79: Add disabledBackground theme color and use for loading states.
+- 2da2dd6: Add support for labels on MetricSeriesChart
+- 564a844: Fixes #450: Adds theming selector to Storybook (and adds an initial dark theme).
+- e0fb6de: Fix #418,#419,#420: Better error states for charts, maps, and compare table.
+- 6a6cda5: Render Skeleton component in charts while data is loading to avoid layout shift.
+- a21eb05: Render map regions in grey while data is loading to avoid layout shifts.
+- ecbedde: implement import sorting
+- Updated dependencies [d69f828]
+- Updated dependencies [ecbedde]
+  - @actnowcoalition/metrics@0.4.0
+  - @actnowcoalition/number-format@0.1.2
+  - @actnowcoalition/regions@0.1.2
+
 ## 0.8.2
 
 ### Patch Changes